--- bonsai/groth16/Cargo.toml
@@ -1,22 +1,0 @@
-[package]
-name = "bonsai-groth16-verifier"
-description = "A library to verify Groth16 proofs computed over the BN_254 curve."
-version = "0.5.0"
-edition = { workspace = true }
-license = { workspace = true }
-homepage = { workspace = true }
-repository = { workspace = true }
-
-[dependencies]
-anyhow = "1"
-ark-bn254 = "0.4"
-ark-groth16 = "0.4"
-ark-serialize = "0.4"
-ethereum-types = "0.14"
-hex = "0.4"
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-sha2 = { version = "^0.10.6" }
-
-[dev-dependencies]
-risc0-zkvm = { workspace = true, default-features = false }
