--- bonsai/groth16/src/fixtures.rs
@@ -1,160 +1,0 @@
-// Copyright 2023 RISC Zero, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-pub(crate) const RISC0_RECEIPT_META_DIGEST: &str =
-    "ff2b40ef5fc5f4a7dac43eb214d61dcf6665f7bedb6cfda244808318e066f656";
-pub(crate) const RISC0_GROTH16_SEAL: &str = r#"
-    {
-        "a":[[26,63,155,211,133,192,185,234,51,172,152,49,113,248,13,45,155,140,75,98,171,225,72,44,133,246,88,199,37,103,28,56],
-            [22,46,102,79,175,35,89,153,55,78,200,143,26,196,209,62,247,200,136,247,101,65,101,157,59,33,20,91,191,43,246,84]],
-        "b":[[[20,0,192,47,211,1,254,55,118,229,52,232,89,161,51,100,224,242,246,5,106,190,188,113,187,230,100,7,255,70,192,153],
-            [9,206,194,232,36,249,103,125,57,1,189,209,245,133,230,79,219,98,176,253,221,160,20,78,189,142,46,52,171,1,162,203]],
-            [[32,70,252,38,183,118,240,156,230,16,28,10,122,111,184,65,239,158,193,102,94,156,5,56,24,236,174,103,160,172,89,109],
-            [4,151,70,37,225,160,65,7,230,206,172,247,41,192,58,149,244,201,31,171,7,137,232,107,29,143,227,49,96,29,198,35]]],
-        "c":[[20,112,216,216,113,186,224,139,117,16,31,102,238,29,76,150,58,216,254,75,147,172,89,147,216,138,41,130,172,106,148,8],
-            [23,250,144,90,188,98,158,62,214,76,108,236,158,31,125,183,20,200,101,254,212,15,32,174,120,230,219,15,71,206,189,55]],
-        "public":[[0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,104,228,45,139,61,220,73,159,78,23,153,167,103,5,42,179],
-            [0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,56,2,104,79,22,69,224,160,40,88,91,4,69,211,146,49],
-            [0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,207,29,214,20,178,62,196,218,167,244,197,95,239,64,43,255],
-            [0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,86,246,102,224,24,131,128,68,162,253,108,219,190,247,101,102]]
-    }
-    "#;
-
-pub(crate) const CIRCOM_VERIFICATION_KEY: &str = r#"
-    {
-        "protocol": "groth16",
-        "curve": "bn128",
-        "nPublic": 1,
-        "vk_alpha_1": [
-        "1294134766316609703328581643861691998063901679593305122518960283123018706388",
-        "13333629383043588737044454681202570079155905422740155054898346012606076806713",
-        "1"
-        ],
-        "vk_beta_2": [
-        [
-        "2173330313723596358484167553880140545051512882245565043987444676076276437843",
-        "17664927106745560489997587182635122110932281433243608150300401610335045630458"
-        ],
-        [
-        "15273531101849588270786039343703563036519656806292651941045419058100734479928",
-        "5906890440295795612829674167362972238653435457353882556276325798552943068201"
-        ],
-        [
-        "1",
-        "0"
-        ]
-        ],
-        "vk_gamma_2": [
-        [
-        "10857046999023057135944570762232829481370756359578518086990519993285655852781",
-        "11559732032986387107991004021392285783925812861821192530917403151452391805634"
-        ],
-        [
-        "8495653923123431417604973247489272438418190587263600148770280649306958101930",
-        "4082367875863433681332203403145435568316851327593401208105741076214120093531"
-        ],
-        [
-        "1",
-        "0"
-        ]
-        ],
-        "vk_delta_2": [
-        [
-        "18082335820320067675049162254051449653127391848352997939790860074257698080107",
-        "8330577861444131504217321247245855407953761241369242366142989304032525780907"
-        ],
-        [
-        "17303423980605275724415088817235493141378511193276153617545225405070114888674",
-        "14329686539600445325529176452626235089284148901536698629845437848687632586506"
-        ],
-        [
-        "1",
-        "0"
-        ]
-        ],
-        "vk_alphabeta_12": [
-        [
-        [
-            "21597631232807937363539811467397773006510227572521934676321553463646334198635",
-            "262163796566031525966924304077669698911462791938684055481358366761190909624"
-        ],
-        [
-            "7906541510069809568866569458625474906165138266731006158097677153173003081190",
-            "6033731974653073317939840745456215697935806048520129111479696325287019924880"
-        ],
-        [
-            "14704987171684462743284913958358496425592435250893903733996815280116183837956",
-            "11976893335360452767634479785443059483596766884568778627130863225715341853664"
-        ]
-        ],
-        [
-        [
-            "12328097080442051249349425344337187894102839822992588206855395089786926203816",
-            "13682208775939290403599679510439179899909912951037259533145887567028127550386"
-        ],
-        [
-            "21192833402016971123221885086549612170051010389337807472438934720324822965947",
-            "13562414185694763175024854871060329561479364355902009699411281367056182859582"
-        ],
-        [
-            "19521540372565909644039072005218101866465290490181239648233003077758316514534",
-            "14972591569740303137698557285367668726475164123365050189180689552096060582998"
-        ]
-        ]
-        ],
-        "IC": [
-        [
-        "14881188593619314262120916669096182039078823054228847940501571078734139590733",
-        "14154402986581165757157012590900333439821186463176177723513413360706693112432",
-        "1"
-        ],
-        [
-        "12590475535581033066201434982368662557531886044597804777316719198629101964198",
-        "15378991198052714418783412681738830395150582056324300616272352953924768221974",
-        "1"
-        ]
-        ]
-    }
-    "#;
-pub(crate) const CIRCOM_PROOF: &str = r#"
-    {
-        "pi_a": [
-        "19752044163435112998099796779947263139365269296294968520404327719124263547111",
-        "11069769267857023583069178672374572453291648685282843843698422556496935187114",
-        "1"
-        ],
-        "pi_b": [
-        [
-        "10648747807246846520146780919185052825636963110330658206295040747407885055071",
-        "12804372218404923567755746304221068640275041956837635530943827697901769703079"
-        ],
-        [
-        "2503338810872511988681832059415719063350505376876347903054293313634087665155",
-        "9633905142041006786673594506047895273339766343254274246797495142581149020665"
-        ],
-        [
-        "1",
-        "0"
-        ]
-        ],
-        "pi_c": [
-        "3377589055768505200338103068502385766692581078477457038865468586522780813958",
-        "3539307538774736362004944548122522044958136460057956047632676706584864343097",
-        "1"
-        ],
-        "protocol": "groth16",
-        "curve": "bn128"
-    }
-    "#;
-pub(crate) const CIRCOM_PUBLIC: &str = r#"["33"]"#;
