+++ bonsai/sdk/Cargo.toml
@@ -9,7 +9,6 @@ repository = { workspace = true }
 
 [dependencies]
 anyhow = "1.0"
-bonsai-groth16-verifier = { workspace = true }
 reqwest = { version = "0.11", features = ["json", "blocking"] }
 serde = { version = "1.0", features = ["derive"] }
 thiserror = "1.0"
