+++ examples/groth16-verifier/methods/guest/Cargo.toml
@@ -6,8 +6,4 @@ edition = "2021"
 [workspace]
 
 [dependencies]
-bonsai-groth16-verifier = { path = "../../../../bonsai/groth16" }
-risc0-zkvm = { path = "../../../../risc0/zkvm", default-features = false, features = [ "std" ] }
-
-[patch.crates-io]
-sha2 = { git = "https://github.com/risc0/RustCrypto-hashes", tag = "sha2-v0.10.6-risczero.0" }
+risc0-zkvm = { path = "../../../../risc0/zkvm", default-features = false }
