+++ examples/groth16-verifier/methods/guest/src/main.rs
@@ -1,12 +1,16 @@
+#![no_std]
 #![no_main]
 
-use bonsai_groth16_verifier::Groth16;
-use risc0_zkvm::guest::env;
+use risc0_zkvm::{
+    groth16::Groth16,
+    guest::env,
+    sha::{Impl, Sha256},
+};
 
 risc0_zkvm::guest::entry!(main);
 
 pub fn main() {
     let proof: Groth16 = env::read();
     proof.verify().unwrap();
-    env::commit(&proof.digest());
+    env::commit(Impl::hash_bytes(&proof.to_vec()));
 }
