+++ risc0/zkvm/Cargo.toml
@@ -34,6 +34,9 @@ protoc-prebuilt = { git = "https://github.com/risc0/protoc-prebuilt.git", rev = 
 
 [dependencies]
 anyhow = { version = "1.0", default-features = false }
+ark-bn254 = { version = "=0.4.0" }
+ark-groth16 = { version = "=0.4.0", default-features = false }
+ark-serialize = { version = "=0.4.2", default-features = false }
 bytemuck = { version = "1.13", features = ["extern_crate_alloc"] }
 cfg-if = "1.0"
 getrandom = { version = "0.2", features = ["custom"] }
@@ -61,6 +64,7 @@ addr2line = { version = "0.21", features = ["rustc-demangle"], optional = true }
 bincode = { version = "1.3", optional = true }
 bonsai-sdk = { workspace = true, optional = true }
 bytes = { version = "1.4", features = ["serde"], optional = true }
+ethereum-types = { version = "0.14", default_features = false }
 rayon = { version = "1.5", optional = true }
 risc0-circuit-recursion = { workspace = true }
 risc0-circuit-rv32im = { workspace = true }
@@ -79,7 +83,7 @@ crypto-bigint = { version = "0.5", default-features = false, features = [
 num-derive = "0.4"
 num-traits = { version = "0.2", default-features = false, optional = true }
 prost = { version = "0.12", optional = true }
-sha2 = { version = "0.10", optional = true }
+sha2 = { version = "0.10" }
 tempfile = { version = "3", optional = true }
 tracing = { version = "0.1", default-features = false, features = [
   "attributes",
@@ -166,7 +170,6 @@ prove = [
   "dep:prost-build",
   "dep:protoc-prebuilt",
   "dep:rayon",
-  "dep:sha2",
   "dep:tempfile",
   "dep:typetag",
   "risc0-circuit-recursion/prove",
