+++ risc0/zkvm/src/host/api/convert.rs
@@ -22,7 +22,7 @@ use risc0_zkp::core::digest::Digest;
 use super::{malformed_err, path_to_string, pb, Asset, AssetRequest, Binary, BinaryKind};
 use crate::{
     host::{
-        receipt::{CompositeReceipt, InnerReceipt, SegmentReceipt},
+        receipt::{CompositeReceipt, InnerReceipt, SegmentReceipt, SnarkReceipt},
         recursion::SuccinctReceipt,
     },
     receipt_metadata::{Assumptions, MaybePruned, Output},
@@ -36,6 +36,7 @@ mod ver {
     pub const RECEIPT: CompatVersion = CompatVersion { value: 1 };
     pub const SEGMENT_RECEIPT: CompatVersion = CompatVersion { value: 1 };
     pub const SUCCINCT_RECEIPT: CompatVersion = CompatVersion { value: 1 };
+    pub const SNARK_RECEIPT: CompatVersion = CompatVersion { value: 1 };
 }
 
 impl TryFrom<AssetRequest> for pb::api::AssetRequest {
@@ -381,6 +382,32 @@ impl TryFrom<pb::core::SegmentReceipt> for SegmentReceipt {
     }
 }
 
+impl From<SnarkReceipt> for pb::core::SnarkReceipt {
+    fn from(value: SnarkReceipt) -> Self {
+        Self {
+            version: Some(ver::SNARK_RECEIPT),
+            seal: value.seal,
+            meta: Some(value.meta.into()),
+        }
+    }
+}
+
+impl TryFrom<pb::core::SnarkReceipt> for SnarkReceipt {
+    type Error = anyhow::Error;
+
+    fn try_from(value: pb::core::SnarkReceipt) -> Result<Self> {
+        let version = value.version.ok_or(malformed_err())?.value;
+        if version > ver::SNARK_RECEIPT.value {
+            bail!("Incompatible SnarkReceipt version: {version}");
+        }
+
+        Ok(Self {
+            seal: value.seal,
+            meta: value.meta.ok_or(malformed_err())?.try_into()?,
+        })
+    }
+}
+
 impl From<SuccinctReceipt> for pb::core::SuccinctReceipt {
     fn from(value: SuccinctReceipt) -> Self {
         Self {
@@ -431,6 +458,7 @@ impl From<InnerReceipt> for pb::core::InnerReceipt {
                         metadata: Some(metadata.into()),
                     })
                 }
+                InnerReceipt::Snark(inner) => pb::core::inner_receipt::Kind::Snark(inner.into()),
             }),
         }
     }
@@ -442,6 +470,7 @@ impl TryFrom<pb::core::InnerReceipt> for InnerReceipt {
     fn try_from(value: pb::core::InnerReceipt) -> Result<Self> {
         Ok(match value.kind.ok_or(malformed_err())? {
             pb::core::inner_receipt::Kind::Composite(inner) => Self::Composite(inner.try_into()?),
+            pb::core::inner_receipt::Kind::Snark(inner) => Self::Snark(inner.try_into()?),
             pb::core::inner_receipt::Kind::Succinct(inner) => Self::Succinct(inner.try_into()?),
             pb::core::inner_receipt::Kind::Fake(inner) => Self::Fake {
                 metadata: inner.metadata.ok_or(malformed_err())?.try_into()?,
